Add files via upload
Buying a car is a big deal -- and saving money is a crucial aspect of any new - or used - car purchase. With that in mind, we're listing a few reasons why a certified pre-owned (or CPO) vehicle may be right for you. First, a certified pre-owned car can be better than a traditional used vehicle because regular used cars often come with a lot of unknowns -- and that could cost you a lot of money. But buying a certified pre-owned vehicle ensures you aren't saddling yourself with any major systems issues or hidden defects